Move convnext to prototype.
